requirements.txt
gradio==3.35.2
matplotlib==3.7.1
numpy==1.24.3
opencv_python==4.7.0.72
Pillow==9.5.0
